requirements.txt
pyautogui
pyperclip
selenium
webdriver-manager
langchain
langchain-google-genai
python-dotenv
PyQt5
pillow
